requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.2
certifi==2025.4.26
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.18
deprecation==2.1.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.12
frozenlist==1.6.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.3
limits==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.3
ordered-set==4.1.0
packaging==25.0
pluggy==1.5.0
postgrest==1.0.1
prompt_toolkit==3.0.51
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
realtime==2.4.2
redis==5.2.1
rich==13.9.4
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
